% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainFlip.R
\name{brainFlip}
\alias{brainFlip}
\title{True Discovery Guarantee for Brain Imaging}
\usage{
sumBrain.internal(copes, mask, clusters, thr, alternative, alpha, B, seed, truncFrom, truncTo,
                  pvalues, type, r, squares, rand, nMax, silent)
}
\arguments{
\item{copes}{list of 3D numeric arrays (contrasts maps for each subject).}

\item{mask}{3D logical array, where \code{TRUE} values correspond to voxels inside the brain, or character for a Nifti file name.}

\item{alternative}{direction of the alternative hypothesis (\code{greater}, \code{lower}, \code{two.sided}).}

\item{alpha}{significance level.}

\item{B}{number of permutations, including the identity.}

\item{seed}{seed.}

\item{truncFrom}{truncation parameter: values less extreme than \code{truncFrom} are truncated.
If \code{NULL}, statistics are not truncated.}

\item{truncTo}{truncation parameter: truncated values are set to \code{truncTo}.
If \code{NULL}, statistics are not truncated.}

\item{pvalues}{logical, \code{TRUE} to use p-values, \code{FALSE} to use t-scores.}

\item{type}{transformation of p-values (\code{edgington}, \code{fisher}, \code{pearson}, \code{liptak},
\code{cauchy}, \code{vovk.wang})}

\item{r}{parameter for Vovk and Wang's p-value transformation.}

\item{squares}{logical, \code{TRUE} to use squared t-scores.}

\item{rand}{logical, \code{TRUE} to compute p-values by permutation distribution.}

\item{clusters}{3D numeric array of cluster indices, or character for a Nifti file name.}

\item{thr}{threshold used to compute the clusters: adjacent t-scores more extreme than \code{thr}
are combined into a single cluster.}

\item{nMax}{maximum number of iterations.}

\item{silent}{logical, \code{FALSE} to print the summary.}
}
\value{
\code{sumBrain.internal} returns a list containing \code{summary} (matrix),
\code{clusters} (3D numeric array of cluster indices), and
\code{TDPmap} (3D numeric array of the true discovery proportions).
The matrix \code{summary} contains, for each cluster:
\itemize{
\item \code{size}: size of the cluster
\item \code{TD}: lower (1-\code{alpha})-confidence bound for the number of true discoveries
\item \code{maxTD}: maximum value of \code{TD} that could be found under convergence of the algorithm
\item \code{TDP}: lower (1-\code{alpha})-confidence bound for the true discovery proportion
\item \code{maxTD}: maximum value of \code{TDP} that could be found under convergence of the algorithm
\item \code{dim1}, \code{dim2}, \code{dim3}: coordinates of the center of mass.
}
}
\description{
Internal function, called in \code{sumBrain} and \code{sumBrain.pvals}.
It determines a true discovery guarantee for fMRI cluster analysis.
}
\details{
A p-value \code{p} is transformed as following.
\itemize{
\item Edgington: \code{-p}
\item Fisher: \code{-log(p)}
\item Pearson: \code{log(1-p)}
\item Liptak: \code{-qnorm(p)}
\item Cauchy: \code{tan(0.5 - p)/p}
\item Vovk and Wang: \code{- sign(r)p^r}
}
An error message is returned if the transformation produces infinite values.

Truncation parameters should be such that \code{truncTo} is not more extreme than \code{truncFrom}.
As Pearson's and Liptak's transformations produce infinite values in 1, for such methods
\code{truncTo} should be strictly smaller than 1.

The significance level \code{alpha} should be in the interval [1/\code{B}, 1).
}
\author{
Anna Vesely.
}
