% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainScores.R
\name{brainScores}
\alias{brainScores}
\title{Matrix of Scores for Brain Imaging}
\usage{
brainScores(copes, mask = NULL, alternative = "two.sided", alpha = 0.05, B = 1000, seed = NULL,
truncFrom = 3.2, truncTo = 0, squares = FALSE)
}
\arguments{
\item{copes}{list of 3D numeric arrays (contrasts maps for each subject).}

\item{mask}{3D logical array, where \code{TRUE} values correspond to voxels inside the brain.}

\item{alternative}{direction of the alternative hypothesis (\code{greater}, \code{lower}, \code{two.sided}).}

\item{alpha}{significance level.}

\item{B}{number of permutations, including the identity.}

\item{seed}{seed.}

\item{truncFrom}{truncation parameter: values less extreme than \code{truncFrom} are truncated.
If \code{NULL}, statistics are not truncated.}

\item{truncTo}{truncation parameter: truncated values are set to \code{truncTo}.
If \code{NULL}, statistics are not truncated.}

\item{squares}{logical, \code{TRUE} to compute squared t-scores.}
}
\value{
\code{brainScores} returns a list containing the matrix of t-scores \code{G},
and the transformed truncation parameters \code{truncFrom} and \code{truncTo}.
In \code{G}, columns correspond to variables, and rows to permutations.
The first permutation is the identity.
}
\description{
This function computes t-scores for different permutations of brain imaging data.
}
\details{
Truncation parameters should be such that \code{truncTo} is not more extreme than \code{truncFrom}.

The significance level \code{alpha} should be in the interval [1/\code{B}, 1).
}
\author{
Anna Vesely.
}
