% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{checkTD}
\alias{checkTD}
\title{Check Candidate TD Value}
\usage{
checkTD(TD, D0, I0, R0, s, f0, k, B, BAB, nMax)
}
\arguments{
\item{TD}{candidate value for the number of true discoveries}

\item{D0}{matrix for the lower bound}

\item{I0}{matrix of indices corresponding to elements in R0}

\item{R0}{matrix for the upper bound}

\item{s}{size of the subset under closed testing}

\item{f0}{number of variables}

\item{k}{integer value that determines the quantile}

\item{B}{number of transformations}

\item{BAB}{number of iterations}

\item{nMax}{maximum number of iterations}
}
\value{
It returns a list containing \code{rej} (\code{TRUE} if no non-rejection was found),
\code{indSizes} (\code{TRUE} if the outcome is unsure) and \code{BAB} (number of iterations).
\code{TD} is a valid confidence lower bound when \code{rej=TRUE} and \code{indSizes=FALSE},
it is not when \code{rej=FALSE} and \code{indSizes=FALSE}, and
the outcome is unsure when \code{rej=TRUE} and \code{indSizes=TRUE}.
}
\description{
Internal function, called in \code{bisectionTD}.
It checks whether a candidate value is a valid lower confidence bound for the number of true discoveries within a subset.
}
\author{
Anna Vesely
}
\keyword{internal}
