% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discoveries.R
\name{discoveries}
\alias{discoveries}
\title{MyHappyFunction: The my happy function}
\usage{
discoveries(obj)
}
\arguments{
\item{x}{numeric number}

\item{...}{other arguments}
}
\description{
A description of MyHappyFunction
}
\details{
A details of MyHappyFunction
}
\examples{
a <- 1
class(a) <- "lm"
MyHappyFunction(a)
}
