% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClusters.R
\name{findClusters}
\alias{findClusters}
\title{Suprathreshold Clusters for Brain Imaging}
\usage{
findClusters(copes, mask = NULL, thr = 3.2, alternative = "two.sided", silent = FALSE)
}
\arguments{
\item{copes}{list of 3D numeric arrays (contrasts maps for each subject).}

\item{mask}{3D logical array, where \code{TRUE} values correspond to voxels inside the brain, or character for a Nifti file name.}

\item{thr}{threshold.}

\item{alternative}{direction of the alternative hypothesis (\code{greater}, \code{lower}, \code{two.sided}).}

\item{silent}{logical, \code{FALSE} to print the number of clusters.}
}
\value{
\code{findClusters} returns a 3D numeric array, with integer values corresponding to clusters,
and 0 to other voxels.
}
\description{
This function determines spatially connected clusters, where t-scores are more extreme than a given threshold.
}
\author{
Anna Vesely.
}
