% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDiscSum.R
\name{findDiscSum}
\alias{findDiscSum}
\title{Find Discoveries Using Summed Statistics}
\usage{
findDiscSum(stats, cvs, ix = NULL, alpha = 0.05)
}
\arguments{
\item{stats}{Numeric vector of \code{m} statistic values. If \code{cvs} is numeric, \code{stats} should be transformed statistics; if \code{cvs} is of type "character", \code{stats} should contain original p-values.}

\item{cvs}{Critical values for sum tests. Can be either a (numeric) \code{m}-vector of critical values for summed transformed statistics considering \code{1:m} hypotheses, or a (character) vector specifying the type of p-value combination test ("hmp.dep", "hmp.ind", "fisher.ind" or "cauchy").}

\item{ix}{Indices of subset of interest \code{S}. If \code{ix = NULL}, it is assumed that the full set is considered.}

\item{alpha}{Significance level.}
}
\value{
\code{findDiscSum} returns a lower (1-\code{alpha})-confidence bound for the number of true discoveries in \code{S}.
}
\description{
Based on a newly developed linear-time shortcut, this function determines lower confidence bounds for the number of true discoveries for a set of interest.
}
\examples{
# generate matrix of p-values for 5 variables and 10 permutations
G <- simData(prop = 0.6, m = 5, B = 10, alpha = 0.4, p = TRUE, seed = 42)

# compute lower confidence bound for the number of true discoveries for the whole set using harmonic mean p-value
res <- findDiscSum(stats = G[1,], cvs = "hmp.ind")

}
\author{
Xu Chen, Jelle Goeman.
}
