% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genePvals.R
\name{genePvals}
\alias{genePvals}
\title{Permutation p-Values for Gene Expression}
\usage{
genePvals(expr, labels, alternative = "two.sided", alpha = 0.05, B = 200, seed = NULL,
          truncFrom = NULL, truncTo = 0.5, type = "vovk.wang", r = 0, rand = FALSE)
}
\arguments{
\item{expr}{matrix where rows correspond to genes, and columns to samples.}

\item{labels}{numeric/character vector with two levels, denoting the population of each sample.}

\item{alternative}{direction of the alternative hypothesis (\code{greater}, \code{lower}, \code{two.sided}).}

\item{alpha}{significance level.}

\item{B}{number of permutations, including the identity.}

\item{seed}{seed.}

\item{truncFrom}{truncation parameter: values greater than \code{truncFrom} are truncated.
If \code{NULL}, it is set to \code{alpha}.}

\item{truncTo}{truncation parameter: truncated values are set to \code{truncTo}.
If \code{NULL}, p-values are not truncated.}

\item{type}{p-value combination among \code{edgington}, \code{fisher}, \code{pearson}, \code{liptak},
\code{cauchy}, \code{harmonic}, \code{vovk.wang} (see details).}

\item{r}{parameter for Vovk and Wang's p-value combination.}

\item{rand}{logical, \code{TRUE} to compute p-values by permutation distribution.}
}
\value{
\code{genePvals} returns an object of class \code{sumGene}, containing
\itemize{
\item \code{statistics}: numeric matrix of p-values, where columns correspond to genes, and rows to permutations.
The first permutation is the identity
\item \code{alpha}: significance level
\item \code{truncFrom}: transformed first truncation parameter
\item \code{truncTo}: transformed second truncation parameter
}
}
\description{
This function computes p-value combinations for different permutations of gene expression data.
A gene's p-value is calculated by performing the two-sample t test
for the null hypothesis that the mean expression value is the same between two populations.
}
\details{
A p-value \code{p} is transformed as following.
\itemize{
\item Edgington: \code{p} (Edgington, 1972)
\item Fisher: \code{-2log(p)} (Fisher, 1925)
\item Pearson: \code{2log(1-p)} (Pearson, 1933)
\item Liptak: \code{qnorm(1-p)} (Liptak, 1958; Stouffer et al., 1949)
\item Cauchy: \code{tan[(0.5-p)pi]} with \code{pi=3.142} (Liu and Xie, 2020)
\item Harmonic mean: \code{1/p} (Wilson, 2019)
\item Vovk and Wang: \code{p^r} (\code{log(p)} for \code{r}=0) (Vovk and Wang, 2020)
}
An error message is returned if the transformation produces infinite values.

For Vovk and Wang, \code{r=0} corresponds to Fisher, and \code{r=-1} to the harmonic mean.

Truncation parameters should be such that \code{truncTo} is not smaller than \code{truncFrom}.
As Pearson's and Liptak's transformations produce infinite values in 1, for such methods
\code{truncTo} should be strictly smaller than 1.

The significance level \code{alpha} should be in the interval [1/\code{B}, 1).
}
\examples{
# simulate 20 samples of 100 genes
set.seed(42)
expr <- matrix(c(rnorm(1000, mean = 0, sd = 10), rnorm(1000, mean = 13, sd = 10)), ncol = 20)
rownames(expr) <- seq(100)
labels <- rep(c(1,2), each = 10)

# simulate pathways
pathways <- lapply(seq(3), FUN = function(x) sample(rownames(expr), 3*x))

# create object of class sumGene
res <- genePvals(expr = expr, labels = labels, alpha = 0.2, seed = 42, type = "liptak")
res
summary(res)

# confidence bound for the number of true discoveries and the TDP within pathways
out <- geneAnalysis(res, pathways = pathways)
out
}
\references{
Goeman J. J. and Solari A. (2011). Multiple testing for exploratory research. Statistical Science, doi: 10.1214/11-STS356.

Vesely A., Finos L., and Goeman J. J. (2023). Permutation-based true discovery guarantee by sum tests. Journal of the Royal Statistical Society, Series B (Statistical Methodology), doi: 10.1093/jrsssb/qkad019.
}
\seealso{
Permutation statistics for gene expression using t scores: \code{\link{geneScores}}

True discovery guarantee for cluster analysis: \code{\link{geneAnalysis}}
}
\author{
Anna Vesely.
}
