% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{permMin}
\alias{permMin}
\title{Compare Vector Elements to Value}
\usage{
permMin(X, B, truncTo)
}
\arguments{
\item{X}{numeric vector}

\item{B}{length of \code{X}}

\item{truncTo}{numeric value}
}
\value{
It returns \code{TRUE} if all elements of \code{X}, excluding the first one, are equal to \code{truncTo}, and returns \code{FALSE} otherwise.
}
\description{
Internal function, called in \code{sumCt.internal}.
It determines whether all the elements of a vector, eventually except the first one, are equal to a given value.
}
\author{
Anna Vesely
}
\keyword{internal}
