% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pimaAnalysis.R
\name{pimaAnalysis}
\alias{pimaAnalysis}
\title{True discovery guarantee in multiverse analysis}
\usage{
pimaAnalysis(pima, by = NULL, type = "sum", r = 0, alpha = 0.05, ...)
}
\arguments{
\item{pima}{an object of class \code{pima}, as obtained from the function \code{\link[pima]{pima}}.}

\item{by}{name of grouping element among \code{Coeff} and \code{Model}. If not specified, all coefficients of interest
in all models are considered together.}

\item{type}{combining function: \code{sum} uses the sum of scores as in \code{\link[sumSome]{sumStats}},
while different p-value combinations are defined as in \code{\link[sumSome]{sumPvals}}.}

\item{r}{parameter for Vovk and Wang's p-value combination.}

\item{alpha}{significance level.}

\item{...}{further parameters of \code{sumStats} or \code{sumPvals} (truncation parameters and maximum number of iterations of the algorithm).}
}
\value{
Returns a data frame containing a summary for each subset:
\itemize{
\item \code{size}: number of considered coefficients
\item \code{TD}: lower (1-\code{alpha})-confidence bound for the number of true discoveries
\item \code{TDP}: lower (1-\code{alpha})-confidence bound for the true discovery proportion
}
}
\description{
This function uses permutation score statistics/p-values to determine a true discovery guarantee for multiverse analysis,
when studying one or more parameters of interest within a multiverse of models.
It computes confidence bounds for the number of true discoveries and the true discovery proportion overall or within different groups.
The bounds are simultaneous over all sets, and remain valid under post-hoc selection.
}
\details{
In the default \code{by = NULL}, the procedure computes lower confidence bounds for the
number/proportion of significant effects (non-null coefficients) among all.
Other inputs of the argument \code{by} return analogous bounds, defined by coefficient (\code{"Coeff"}) or by model (\code{"Model"}).

If truncation parameters are not specified, scores/p-values are not truncated.
}
\examples{
# COPIARE PARTE DELL'ESEMPIO DELLA FUNZIONE pima()
}
\references{
Girardi, Vesely, Lakens, Altoè, Pastore, Calcagnì, Finos (2024). Post-selection Inference in Multiverse Analysis (PIMA): An Inferential Framework Based on the Sign Flipping Score Test. Psychometrika, doi: 10.1007/s11336-024-09973-6.

Vesely, Finos, Goeman (2023). Permutation-based true discovery guarantee by sum tests. Journal of the Royal Statistical Society, Series B (Statistical Methodology), doi: 10.1093/jrsssb/qkad019
}
