% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateMat.R
\name{simulateMat}
\alias{simulateMat}
\title{Simulating Matrix of Statistics}
\usage{
simulateMat(prop, m = 1000, B = 200, rho = 0, n = 50, alpha = 0.05, power = 0.8, pvalues = TRUE,
            rand = FALSE, seed = NULL)
}
\arguments{
\item{prop}{proportion of non-null hypotheses.}

\item{m}{total number of variables.}

\item{B}{number of permutations, including the identity.}

\item{rho}{level of equicorrelation between pairs of variables.}

\item{n}{number of observations.}

\item{alpha}{significance level.}

\item{power}{power of the one-sample t test.}

\item{pvalues}{logical, \code{TRUE} to compute p-values, \code{FALSE} to return t-scores.}

\item{rand}{logical, \code{TRUE} to compute p-values by permutation distribution.}

\item{seed}{seed.}
}
\value{
\code{simulateMat} returns a matrix where the \code{B} rows correspond to permutations (the first is the identity),
and the \code{m} columns correspond to variables.
The matrix contains p-values if \code{pvalues} is \code{TRUE}, and t-scores otherwise.
The first columns (a proportion \code{prop}) correspond to non-null hypotheses.
}
\description{
This function simulates a matrix of statistics that may be used
to test \code{\link{sumSome}} and \code{\link{sumSome.pvals}}.
}
\details{
The function applies the one-sample two-sided t test to a matrix of simulated data,
for \code{B} data permutations.
Data is obtained by simulating \code{n} independent observations from a multivariate normal distribution,
where a proportion \code{prop} of the variables has non-null mean.
This mean is such that the one-sample t test with significance level \code{alpha} has power equal to \code{power}.
Each pair of distinct variables has equicorrelation \code{rho}.
}
\examples{
# simulate 10 x 5 matrix of p-values
# the first 3 variables correspond to non-null hypotheses
G <- simulateMat(prop = 0.6, m = 5, B = 10, alpha = 0.4, seed = 42)

sumSome.pvals(G, S = c(1,2,3), alpha = 0.4, type = "vovk.wang", r = -1)
}
\author{
Anna Vesely.
}
