% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePvals.R
\name{simulatePvals}
\alias{simulatePvals}
\title{Simulations for p-Value Combinations}
\usage{
simulatePvals(type = "fisher", r = 1, rho = c(0,0.33,0.66,0.99), prop = c(0.1,0.2,0.3,0.4), m = 1000,
              B = 200, n = 50, alpha = 0.05, power = 0.8, rand = FALSE, truncFrom = alpha, truncTo = 1,
              nMax = 10000, nSim = 1000, silent = FALSE)
}
\arguments{
\item{type}{transformation of p-values (\code{edgington}, \code{fisher}, \code{pearson}, \code{liptak},
\code{cauchy}, \code{vovk.wang})}

\item{r}{parameter for Vovk and Wang's p-value transformation.}

\item{rho}{vector of levels of equicorrelation between pairs of variables.}

\item{prop}{vector of proportions of non-null hypotheses.}

\item{m}{total number of variables.}

\item{B}{number of permutations, including the identity.}

\item{n}{number of observations.}

\item{alpha}{significance level.}

\item{power}{power of the one-sample t test.}

\item{rand}{logical, \code{TRUE} to compute p-values by permutation distribution.}

\item{truncFrom}{truncation parameter: values greater than \code{truncFrom} are truncated.
If \code{NULL}, p-values are not truncated.}

\item{truncTo}{truncation parameter: truncated values are set to \code{truncTo}.
If \code{NULL}, p-values are not truncated.}

\item{nMax}{maximum number of iterations.}

\item{nSim}{number of simulations per scenario.}

\item{silent}{logical, \code{FALSE} to print the summary.}
}
\value{
\code{simulatePvals} returns two lists, \code{results} and \code{summary}.
The list \code{results} contains, for each value of \code{prop}, the matrices
\code{TD}, \code{iterations} and \code{convergence}, which describe the results
for each value of \code{rho} and seed.
The list summary contains the matrices
\itemize{
\item \code{TDP}: matrix of the mean true discovery proportion for scenario
\item \code{iterations}: matrix of the mean number of iterations for scenario
\item \code{convergence}: logical, \code{TRUE} if the algorithm converges in all the simulations
}
}
\description{
This function describes the behavior of a p-value combination in a given setting
through simulations.
}
\details{
For each scenario, the function generates
\code{nSim} matrices of simulated p-values through \code{\link{simulateMat}},
with seed equal to 1,...,\code{nSim}.
Then it determines a lower confidence bound for the number of true discoveries
within the set of non-null hypotheses.

The significance level \code{alpha} should be in the interval [1/\code{B}, 1).

Truncation parameters should be such that \code{truncTo} is not smaller than \code{truncFrom}.

A p-value \code{p} is transformed as following.
\itemize{
\item Edgington: \code{-p}
\item Fisher: \code{-log(p)}
\item Pearson: \code{log(1-p)}
\item Liptak: \code{-qnorm(p)}
\item Cauchy: \code{tan(0.5 - p)/p}
\item Vovk and Wang: \code{- sign(r)p^r}
}

Pearson's and Liptak's transformations produce infinite values in \code{1}.
For such transformations, \code{truncTo} is coerced to be not greater than \code{1 -  .Machine$double.eps}.
}
\examples{
# 16 scenarions, 10 simulations each
sim <- simulatePvals(type="vovk.wang", r=-1, rho=c(0,0.33,0.66,0.99), prop = c(0.1,0.2,0.3,0.4), nSim=10)
}
\author{
Anna Vesely.
}
