% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.internal.R
\name{sum.internal}
\alias{sum.internal}
\title{True Discovery Guarantee (General Case)}
\usage{
sum.internal(G, S, alpha, truncFrom, truncTo, nMax)
}
\arguments{
\item{G}{numeric matrix of statistics, where columns correspond to variables, and rows to data transformations (e.g. permutations).
The first transformation is the identity. Extreme values are the greatest.}

\item{S}{vector of indices for the variables of interest.}

\item{alpha}{significance level.}

\item{truncFrom}{truncation parameter: values less extreme than \code{truncFrom} are truncated.
If \code{NULL}, statistics are not truncated.}

\item{truncTo}{truncation parameter: truncated values are set to \code{truncTo}.
If \code{NULL}, statistics are not truncated.}

\item{nMax}{maximum number of iterations.}
}
\value{
\code{sum.internal} returns a list containing \code{summary} (vector) and \code{iterations} (number of iterations).
\code{summary} contains:
\itemize{
\item \code{size}: size of \code{S}
\item \code{TD}: lower (1-\code{alpha})-confidence bound for the number of true discoveries in \code{S}
\item \code{maxTD}: maximum value of \code{TD} that could be found under convergence of the algorithm
\item \code{TDP}: lower (1-\code{alpha})-confidence bound for the true discovery proportion in \code{S}
\item \code{maxTD}: maximum value of \code{TDP} that could be found under convergence of the algorithm
}
}
\description{
Internal function, called in \code{sumSome}, \code{sumSome.pvals} and \code{sumBrain.internal}.
It determines a lower confidence bound for the number of true discoveries within a set of interest.
The bound remains valid under post-hoc selection.
}
\details{
The significance level \code{alpha} should be in the interval [1/\code{B}, 1).
}
\author{
Anna Vesely.
}
\keyword{internal}
