% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumBrain_class.R
\name{sumBrain}
\alias{sumBrain}
\title{sumBrain Class}
\usage{
sumBrain(G, mask, alpha, truncFrom, truncTo)
}
\arguments{
\item{G}{numeric matrix of statistics, where columns correspond to voxels inside the brain, and rows to permutations.
The first permutation is the identity.}

\item{mask}{3D logical array, where \code{TRUE} values correspond to voxels inside the brain.}

\item{alpha}{significance level.}

\item{truncFrom}{truncation parameter: values less extreme than \code{truncFrom} are truncated.
If \code{NULL}, statistics are not truncated.}

\item{truncTo}{truncation parameter: truncated values are set to \code{truncTo}.
If \code{NULL}, statistics are not truncated.}
}
\value{
\code{sumBrain} returns an object of class \code{sumBrain}, containing
\code{statistics} (matrix \code{G}), \code{mask}, \code{alpha}, \code{truncFrom} and \code{truncTo}.
}
\description{
Internal function.
It defines a \code{sumBrain} object, storing information on the test statistics
for different permutations of brain imaging data.
}
\author{
Anna Vesely.
}
\keyword{Internal}
