% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumSome_class.R
\name{sumSome}
\alias{sumSome}
\title{sumSome Class}
\usage{
sumSome(total, size, alpha, TD, maxTD, iterations)
}
\arguments{
\item{total}{total number of variables.}

\item{size}{size of the set of interest.}

\item{alpha}{significance level.}

\item{TD}{lower (1-\code{alpha})-confidence bound for the number of true discoveries in the set.}

\item{maxTD}{maximum value of \code{TD} that could be found under convergence of the algorithm.}

\item{iterations}{number of iterations of the algorithm.}
}
\value{
\code{sumSome} returns an object of class \code{sumSome}, containing
\code{total}, \code{size}, \code{alpha}, \code{TD}, \code{maxTD} and \code{iterations}.
}
\description{
Internal function.
It defines a \code{sumSome} object, storing information on the closed testing procedure on a set of interest.
}
\author{
Anna Vesely.
}
\keyword{Internal}
