% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumSome_stats.R
\name{sumSome.stats}
\alias{sumSome.stats}
\title{True Discovery Guarantee}
\usage{
sumSome.stats(G, S, alternative = "greater", alpha = 0.05, truncFrom = NULL, truncTo = NULL, nMax = 10000)
}
\arguments{
\item{G}{numeric matrix of statistics, where columns correspond to variables, and rows to data transformations (e.g. permutations).}

\item{S}{vector of indices for the variables of interest.}

\item{alternative}{direction of the alternative hypothesis (\code{greater}, \code{lower}, \code{two.sided}).}

\item{alpha}{significance level.}

\item{truncFrom}{truncation parameter: values less extreme than \code{truncFrom} are truncated.
If \code{NULL}, statistics are not truncated.}

\item{truncTo}{truncation parameter: truncated values are set to \code{truncTo}.
If \code{NULL}, statistics are not truncated.}

\item{nMax}{maximum number of iterations.}
}
\value{
\code{sumSome.stats} returns a list containing \code{summary} (vector) and \code{iterations} (number of iterations).
The vector \code{summary} contains
\itemize{
\item \code{size}: size of \code{S}
\item \code{TD}: lower (1-\code{alpha})-confidence bound for the number of true discoveries in \code{S}
\item \code{maxTD}: maximum value of \code{TD} that could be found under convergence of the algorithm
\item \code{TDP}: lower (1-\code{alpha})-confidence bound for the true discovery proportion in \code{S}
\item \code{maxTD}: maximum value of \code{TDP} that could be found under convergence of the algorithm
}
}
\description{
This function determines a lower confidence bound for the number of true discoveries within a set of interest.
The bound remains valid under post-hoc selection.
}
\details{
The significance level \code{alpha} should be in the interval [1/\code{B}, 1).

Truncation parameters should be such that \code{truncTo} is not more extreme than \code{truncFrom}.
}
\examples{
G <- matrix(
 c(6,5,4,1,1,
  1,2,1,0,4,
  8,3,0,2,1,
  8,1,0,1,0,
  0,6,1,1,2,
  7,0,1,2,1),
 ncol=5, byrow=TRUE)
S <- c(1,2)
sumSome.stats(G, S, alternative = "greater", alpha = 0.4, truncFrom = 2, truncTo = 0)
}
\author{
Anna Vesely.
}
