% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainFlip.R
\name{brainFlip}
\alias{brainFlip}
\title{Permutation Statistics for Brain Imaging}
\usage{
brainFlip(copes, mask, alternative, alpha, B, seed, truncFrom, truncTo, pvalues, type, r, squares, rand)
}
\arguments{
\item{copes}{list of 3D numeric arrays (contrasts maps for each subject).}

\item{mask}{3D logical array, where \code{TRUE} values correspond to voxels inside the brain, or character for a Nifti file name.}

\item{alternative}{direction of the alternative hypothesis (\code{greater}, \code{lower}, \code{two.sided}).}

\item{alpha}{significance level.}

\item{B}{number of permutations, including the identity.}

\item{seed}{seed.}

\item{truncFrom}{truncation parameter: values less extreme than \code{truncFrom} are truncated.
If \code{NULL}, statistics are not truncated.}

\item{truncTo}{truncation parameter: truncated values are set to \code{truncTo}.
If \code{NULL}, statistics are not truncated.}

\item{pvalues}{logical, \code{TRUE} to use p-values, \code{FALSE} to use t-scores.}

\item{type}{p-value combination (\code{edgington}, \code{fisher}, \code{pearson}, \code{liptak},
\code{cauchy}, \code{vovk.wang})}

\item{r}{parameter for Vovk and Wang's p-value transformation.}

\item{squares}{logical, \code{TRUE} to use squared t-scores.}

\item{rand}{logical, \code{TRUE} to compute p-values by permutation distribution.}
}
\value{
\code{brainFlip} returns an object of class \code{sumBrain}, containing
\itemize{
\item \code{statistics}: numeric matrix of statistics, where columns correspond to voxels inside the brain, and rows to permutations.
The first permutation is the identity
\item \code{mask}: 3D logical array, where \code{TRUE} values correspond to voxels inside the brain
\item \code{alpha}: significance level
\item \code{truncFrom}: transformed first truncation parameter
\item \code{truncTo}: transformed second truncation parameter
}
}
\description{
Internal function.
It computes test statistics for different permutations of brain imaging data.
A voxel's statistic is calculated by performing the one-sample t test
for the null hypothesis that its mean contrast over the different subjects is zero.
}
\details{
A p-value \code{p} is transformed as following.
\itemize{
\item Edgington: \code{-p}
\item Fisher: \code{-log(p)}
\item Pearson: \code{log(1-p)}
\item Liptak: \code{-qnorm(p)}
\item Cauchy: \code{tan(0.5 - p)/p}
\item Vovk and Wang: \code{- sign(r)p^r}
}
An error message is returned if the transformation produces infinite values.

Truncation parameters should be such that \code{truncTo} is not more extreme than \code{truncFrom}.
As Pearson's and Liptak's transformations produce infinite values in 1, for such methods
\code{truncTo} should be strictly smaller than 1.

The significance level \code{alpha} should be in the interval [1/\code{B}, 1).
}
\author{
Anna Vesely.
}
\keyword{Internal}
