% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterAnalysis.R
\name{clusterAnalysis}
\alias{clusterAnalysis}
\title{True Discovery Guarantee for Cluster Analysis}
\usage{
clusterAnalysis(sumBrain, clusters, nMax = 50, silent = FALSE)
}
\arguments{
\item{sumBrain}{an object of class sumBrain, as returned by the functions \code{\link{brainScores}} and \code{\link{brainPvals}}.}

\item{clusters}{3D numeric array of cluster indices, or character for a Nifti file name.
If NULL, the whole brain is considered.}

\item{nMax}{maximum number of iterations per cluster.}

\item{silent}{logical, \code{FALSE} to print the summary.}
}
\value{
\code{clusterAnalysis} returns a list containing \code{summary} (matrix) and
\code{TDPmap} (3D numeric array of the true discovery proportions).
The matrix \code{summary} contains, for each cluster,
\itemize{
\item \code{size}: size
\item \code{TD}: lower (1-\code{alpha})-confidence bound for the number of true discoveries
\item \code{maxTD}: maximum value of \code{TD} that could be found under convergence of the algorithm
\item \code{TDP}: lower (1-\code{alpha})-confidence bound for the true discovery proportion
\item \code{maxTD}: maximum value of \code{TDP} that could be found under convergence of the algorithm
\item \code{dim1}, \code{dim2}, \code{dim3}: coordinates of the center of mass.
}
}
\description{
This function determines a true discovery guarantee for fMRI cluster analysis.
It computes confidence bounds for the number of true discoveries and the true discovery proportion
within each cluster. The bounds are simultaneous over all sets, and remain valid under post-hoc selection.
}
\author{
Anna Vesely.
}
