% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneAnalysis.R
\name{geneAnalysis}
\alias{geneAnalysis}
\title{True Discovery Guarantee for Pathway Analysis of Gene Expression Data}
\usage{
geneAnalysis(sumGene, pathways = NULL, nMax = 50, silent = FALSE)
}
\arguments{
\item{sumGene}{an object of class sumGene, as returned by the functions \code{\link{geneScores}} and \code{\link{genePvals}}.}

\item{pathways}{list of character vectors containing gene names (one vector per pathway). If NULL, the whole gene set is considered.}

\item{nMax}{maximum number of iterations per cluster.}

\item{silent}{logical, \code{FALSE} to print a summary of active pathways.}
}
\value{
\code{geneAnalysis} returns a data frame containing, for each pathway,
\itemize{
\item \code{size}: size
\item \code{TD}: lower (1-\code{alpha})-confidence bound for the number of true discoveries
\item \code{maxTD}: maximum value of \code{TD} that could be found under convergence of the algorithm
\item \code{TDP}: lower (1-\code{alpha})-confidence bound for the true discovery proportion
\item \code{maxTD}: maximum value of \code{TDP} that could be found under convergence of the algorithm.
}
}
\description{
This function uses permutation t-statistics/p-values to determine a true discovery guarantee
for gene pathway analysis.
It computes confidence bounds for the number of true discoveries and the true discovery proportion
within each cluster. The bounds are simultaneous over all sets, and remain valid under post-hoc selection.
}
\examples{
# simulate 20 samples of 100 genes
set.seed(42)
expr <- matrix(c(rnorm(1000, mean = 0, sd = 10), rnorm(1000, mean = 13, sd = 10)), ncol = 20)
rownames(expr) <- seq(100)
labels <- rep(c(1,2), each = 10)

# simulate pathways
pathways <- lapply(seq(3), FUN = function(x) sample(rownames(expr), 3*x))

# create object of class sumGene
res <- geneScores(expr = expr, labels = labels, alpha = 0.2, seed = 42)
res
summary(res)

# confidence bound for the number of true discoveries and the TDP within pathways
out <- geneAnalysis(res, pathways = pathways)
out
}
\references{
Goeman, J. J. and Solari, A. (2011). Multiple testing for exploratory research. Statistical Science, 26(4):584-597.

Hemerik, J. and Goeman, J. J. (2018). False discovery proportion estimation by permutations: confidence for significance analysis of microarrays. JRSS B, 80(1):137-155.

Vesely, A., Finos, L., and Goeman, J. J. (2021). Permutation-based true discovery guarantee by sum tests. Pre-print arXiv:2102.11759.
}
\seealso{
Permutation statistics for gene expression: \code{\link{geneScores}}, \code{\link{genePvals}}
}
\author{
Anna Vesely.
}
