% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumBrain.R
\name{sumBrain}
\alias{sumBrain}
\title{True Discovery Guarantee for Brain Imaging with t-Scores}
\usage{
sumBrain(copes, mask = NULL, clusters = NULL, thr = 3.2, alternative = "two.sided", alpha = 0.05,
         B = 1000, seed = NULL, truncFrom = thr, truncTo = 0, squares = FALSE, nMax = 10000,
         silent = FALSE)
}
\arguments{
\item{copes}{list of 3D numeric arrays (contrasts maps for each subject).}

\item{mask}{3D logical array, where \code{TRUE} values correspond to voxels inside the brain.}

\item{clusters}{3D numeric array of cluster indices, or character for a Nifti file name.}

\item{thr}{threshold used to compute the clusters: adjacent t-scores more extreme than \code{thr}
are combined into a single cluster.}

\item{alternative}{direction of the alternative hypothesis (\code{greater}, \code{lower}, \code{two.sided}).}

\item{alpha}{significance level.}

\item{B}{number of permutations, including the identity.}

\item{seed}{seed.}

\item{truncFrom}{truncation parameter: values less extreme than \code{truncFrom} are truncated.
If \code{NULL}, statistics are not truncated.}

\item{truncTo}{truncation parameter: truncated values are set to \code{truncTo}.
If \code{NULL}, statistics are not truncated.}

\item{squares}{logical, \code{TRUE} to use squared t-scores.}

\item{nMax}{maximum number of iterations.}

\item{silent}{logical, \code{FALSE} to print the summary.}
}
\value{
\code{sumBrain} returns a list containing \code{summary} (matrix),
\code{clusters} (3D numeric array of cluster indices), and
\code{TDPmap} (3D numeric array of the true discovery proportions).
The matrix \code{summary} contains, for each cluster:
\itemize{
\item \code{size}: size of the cluster
\item \code{TD}: lower (1-\code{alpha})-confidence bound for the number of true discoveries
\item \code{maxTD}: maximum value of \code{TD} that could be found under convergence of the algorithm
\item \code{TDP}: lower (1-\code{alpha})-confidence bound for the true discovery proportion
\item \code{maxTD}: maximum value of \code{TDP} that could be found under convergence of the algorithm
\item \code{dim1}, \code{dim2}, \code{dim3}: coordinates of the center of mass.
}
}
\description{
This function determines a true discovery guarantee for fMRI cluster analysis, using t-scores.
}
\details{
The significance level \code{alpha} should be in the interval [1/\code{B}, 1).

Truncation parameters should be such that \code{truncTo} is not more extreme than \code{truncFrom}.
}
\examples{
# if needed, install the package fMRIdata from Github
devtools::install_github("angeella/fMRIdata")

library(fMRIdata)
data("Auditory_copes")
data("Auditory_mask")
data("Auditory_clusterTH3_2")

# the following requires some minutes
out <- sumBrain(copes = Auditory_copes, mask = Auditory_mask, clusters = Auditory_clusterTH3_2,
                B = 100, seed = 42, nMax = 30)

# write the TDP map as Nifti file
library(RNifti)
RNifti::writeNifti(out$TDPmap, file = "TDPmap.nii.gz")
}
\author{
Anna Vesely.
}
