\name{sumSome-package}
\alias{sumSome-package}
\docType{package}
\title{
  True Discovery Guarantee by Sum-Based Tests
}
\description{
  \code{sumSome} provides permutation-based true discovery guarantees, using sum-based global statistics 
  (sum of t-scores, p-value combinations etc.). As a main feature, it produces simultaneous lower confidence bounds
  for the number of active voxels in different clusters for fMRI cluster analysis.
}
\author{
Anna Vesely.

Maintainer: Anna Vesely <anna.vesely@phd.unipd.it>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Generic statistics: \code{\link{sumSome}}
  
  p-value combinations: \code{\link{sumSome.pvals}}
  
  Brain imaging with t-scores: \code{\link{sumBrain}}
  
  Brain imaging with p-values: \code{\link{sumBrain.pvals}}
}

