\name{sumSome-package}
\alias{sumSome-package}
\alias{sumSome}
\docType{package}
\title{
  True Discovery Guarantee by Sum-Based Tests
}
\description{
  \code{sumSome} provides permutation-based true discovery guarantees, using sum-based global statistics 
  (sum of t-scores, p-value combinations etc.). As a main feature, it produces simultaneous lower confidence bounds
  for the number of active voxels in different clusters for fMRI cluster analysis.
}
\author{
Anna Vesely.

Maintainer: Anna Vesely <anna.vesely@phd.unipd.it>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  True discovery guarantee: \code{\link{sumStats}}, \code{\link{sumPvals}}
  
  Permutation statistics for brain imaging: \code{\link{brainScores}}, \code{\link{brainPvals}}
  
  True discovery guarantee for cluster analysis: \code{\link{clusterAnalysis}}
  
  Simulate data: \code{\link{simData}}
}

