% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf.R
\name{transf}
\alias{transf}
\title{Statistics Transformation}
\usage{
transf(G, truncFrom, truncTo, option, r)
}
\arguments{
\item{G}{matrix of statistics}

\item{truncFrom}{truncation parameter: values less extreme than \code{truncFrom} are truncated}

\item{truncTo}{truncation parameter: truncated values are set to \code{truncTo}}

\item{option}{direction of the alternative hypothesis (\code{greater}, \code{lower}, \code{two.sided}),
or transformation (\code{squares} for generic statistics,
and \code{edgington}, \code{fisher}, \code{pearson}, \code{liptak}, \code{cauchy}, \code{vovk.wang} for p-values)}

\item{r}{parameter for the Vovk and Wang p-value transformation}
}
\value{
\code{transf} returns a list containing the truncated and transformed matrix \code{G}, and the transformed truncation parameters \code{truncFrom} and \code{truncTo}.
Those are determined so that the most extreme values are always the greatest.
}
\description{
Internal function. It truncates and transforms a matrix of statistics.
}
\details{
The transformation is determined by the \code{option} as following. A generic value \code{x} is transformed to
\code{x} for \code{greater}, \code{-x} for \code{lower}, \code{|x|} for \code{two.sided},
\code{x^2} for \code{squares}, \code{-x} for \code{edgington}, \code{-log(x)} for \code{fisher},
\code{log(1-x)} for \code{pearson}, \code{-qnorm(x)} for \code{liptak}, \code{tan(0.5 - x)/x} for \code{cauchy},
and \code{- sign(r)x^r} for \code{vovk.wang}.

Pearson's and Liptak's transformations produce infinite values in \code{1}.
Hence when the \code{option} is \code{pearson} or \code{liptak}, the truncation parameter \code{truncTo} is coerced to be
not greater than \code{1 -  .Machine$double.eps}.
}
\author{
Anna Vesely
}
